requirements/base.txt
asgiref==3.5.2
Django==4.1
sqlparse==0.4.2
django-environ==0.9.0
django-taggit==3.0.0
django-admin-rangefilter==0.9.0
django-taggit==3.0.0
django-unicorn==0.48.0
django-ckeditor-5==0.2.1